861246: UMLNodes
