+++ ej2-javascript/code-snippet/diagram/umldiagramshapes-cs1/index.ts
@@ -7,13 +7,16 @@ import {
 } from "@syncfusion/ej2-diagrams";
 let node: NodeModel = {
     id: "Patient",
+    style: {
+      fill: '#26A0DA',
+    },
     //Position of the node
     offsetX: 200,
     offsetY: 200,
     shape: {
       type: "UmlClassifier",
       //Define class object
-      class: {
+      classShape: {
         name: "Patient",
         //Define class attributes
         attributes: [{ name: "accepted", type: "Date" }],
