+++ ej2-javascript/code-snippet/diagram/umldiagramshapes-cs12/index.css
@@ -15,4 +15,5 @@
 
 #element {
     display: block;
+    height: 350px;
 }
