+++ ej2-javascript/code-snippet/diagram/umldiagramshapes-cs12/js/index.html
@@ -5,6 +5,7 @@
     <meta name="description" content="Typescript UI Controls">
     <meta name="author" content="Syncfusion">
     <link href="index.css" rel="stylesheet">
+    <link href="https://cdn.syncfusion.com/ej2/20.3.56/ej2-popups/styles/material.css" rel="stylesheet">
     
     <link href="https://cdn.syncfusion.com/ej2/20.3.56/ej2-base/styles/material.css" rel="stylesheet">
     <link href="https://cdn.syncfusion.com/ej2/20.3.56/ej2-buttons/styles/material.css" rel="stylesheet">
@@ -15,12 +16,12 @@
     
 <script src="https://cdn.syncfusion.com/ej2/20.4.38/dist/ej2.min.js" type="text/javascript"></script>
 </head>
-
 <body>
     
     <div id="container">
         <div id="element"></div>
     </div>
+
 
 <script>
 var ele = document.getElementById('container');
