+++ ej2-javascript/code-snippet/diagram/umldiagramshapes-cs12/systemjs.config.js
