+++ ej2-javascript/code-snippet/diagram/umldiagramshapes-cs12/ts/index.html
@@ -8,6 +8,7 @@
     <meta name="description" content="Typescript UI Controls" />
     <meta name="author" content="Syncfusion" />
     <link href="index.css" rel="stylesheet" />
+    <link href="https://cdn.syncfusion.com/ej2/20.3.56/ej2-popups/styles/material.css" rel="stylesheet" />
     <script src="https://cdnjs.cloudflare.com/ajax/libs/systemjs/0.19.38/system.js"></script>
     <link href="https://cdn.syncfusion.com/ej2/20.3.56/ej2-base/styles/material.css" rel="stylesheet" />
     <link href="https://cdn.syncfusion.com/ej2/20.3.56/ej2-buttons/styles/material.css" rel="stylesheet" />
@@ -17,7 +18,6 @@
     <link href="https://cdn.syncfusion.com/ej2/20.3.56/ej2-navigations/styles/fabric.css" rel="stylesheet" />
     <script src="systemjs.config.js"></script>
 </head>
-
 <body>
     <div id='loader'>Loading....</div>
     <div id='container'>
