+++ ej2-javascript/code-snippet/diagram/umldiagramshapes-cs2/index.js
@@ -6,6 +6,9 @@ var node = {
     id: 'node',
         offsetX: 400,
         offsetY: 300,
+        style: {
+            fill: '#26A0DA',
+        },
         shape: {
             type: 'UmlClassifier',
             interfaceShape: {
