+++ ej2-javascript/code-snippet/diagram/umldiagramshapes-cs2/index.ts
@@ -10,10 +10,13 @@ let node: NodeModel = {
     //Position of the node
     offsetX: 200,
     offsetY: 200,
+    style: {
+      fill: '#26A0DA',
+    },
     shape: {
       type: "UmlClassifier",
       //Define interface object
-      interface: {
+      interfaceShape: {
         name: "Patient",
         //Define interface attributes
         attributes: [{ name: "owner", type: "String[*]" }],
