+++ ej2-javascript/code-snippet/diagram/umldiagramshapes-cs3/index.ts
@@ -9,10 +9,13 @@ let node: NodeModel = {
     id: "Patient",
     offsetX: 200,
     offsetY: 200,
+    style: {
+      fill: '#26A0DA',
+    },
     shape: {
       type: "UmlClassifier",
       //Define enumeration object
-      enumeration: {
+      enumerationShape: {
         name: "AccountType",
         //set the members of enumeration
         members: [
